add blur config
OptiFine changes the path to chat, which means the default blur exclusion does not work. Having the game blur when opening chat, while might look cool, is not intended and may be annoying or distracting. I've only seen one user complain about it but it still needs to be changed. PR because I forgor if you have to do anything more than this :skull: