revert #350 (more below)
as the original suggester of BZNotifier there is no reason to give in to the 3478 duplicate suggestions for BZNotifier until we have a proper, multi-person discussion about whether it deserves to be in skyclient.

if memory serves correctly, the original reason barring BZNotifier from being included in the skyclient mod library was because the author has a history of paywalling most of his other mods. evidence below, if you think the discord messages are falsified then the search bar in discord is your friend:

— https://canary.discord.com/channels/780181693100982273/887818760126345246/975425447334645820
— https://canary.discord.com/channels/780181693100982273/887818760126345246/975425579505578025
— https://canary.discord.com/channels/780181693100982273/887818760126345246/975425625760358461
— https://canary.discord.com/channels/780181693100982273/887818760126345246/975425669024591892
— https://canary.discord.com/channels/780181693100982273/887818760126345246/975425702377717840

also, have we even gotten direct permission from meyi yet? even with all the paywall history aside?